Extract EventFields to a separate file
