+++ test/authorFields.ts
@@ -1,0 +1,10 @@
+// eslint-disable-next-line import/prefer-default-export
+export const authorFields = `
+    {
+        id,
+        username,
+        name,
+        surname,
+        email,
+    }
+`;
