+++ test/givenAwayEvent.graphql.test.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import supertest from 'supertest';
 import validate from './validators/givenAwayEvent.interface.validator';
-import { authorFields } from './event.graphql.test';
+import { authorFields } from './authorFields';
 
 require('dotenv').config({ path: './test/.env' });
